+++ src/instances/Cat.java
@@ -1,9 +1,9 @@
 package instances;
 
-import jumpInstance.Universal;
-import jumpInstance.Instance;
+import Main.Universal;
+import Main.InstancesParent;
 
-public class Cat extends Instance implements Universal {
+public class Cat extends InstancesParent implements Universal {
 int jump;
 
     public Cat(int jump, int run, String name)
