+++ src/instances/Robot.java
@@ -1,9 +1,9 @@
 package instances;
 
-import jumpInstance.Universal;
-import jumpInstance.Instance;
+import Main.Universal;
+import Main.InstancesParent;
 
-public class Robot extends Instance implements Universal {
+public class Robot extends InstancesParent implements Universal {
     int jump;
 
     public Robot(int jump, int run, String name)
